Force doubles in the LLL-reduced array in numpy.
Former-commit-id: 70cffb53ff5b15515da8dd8ac18f74721f2c3b48

+++ qpms/qpms_c.pyx
@@ -619,7 +619,7 @@ def lll_reduce(basis):
     and d is the dimensionality of the space into which the lattice
     is embedded.
     """
-    basis = np.array(basis, copy=True, order='C')
+    basis = np.array(basis, copy=True, order='C', dtype=np.double)
     if len(basis.shape) != 2:
         raise ValueError("Expected two-dimensional array (got %d-dimensional)"
                 % len(basis.shape))
